Downloader: close URL connection after download (#2175)
The problem was that two connections for the Wine versions json are opened. The first to check if an update is available (Downloader#isUpdateAvailable()), the second to actually download the file (Downloader#get()). If e.g. Downloader#isUpdateAvailable() does not open a connection but returns true immediately, the download in Downloader#get() starts directly.

I'm not completely sure what happens in the background but my assumption is that the underlying socket is closed by the disconnect such that the next connection can setup a new socket immediately.