merge-me action: check only CI (#2603)
As described in ridedott/merge-me-action#1212, the action runs as soon as any workflow in workflows completed. This means that it usually ran after linting the Java files and thus ignored if the CI failed.

Ignoring the linting workflow is not optimal. However, the current behavior is way worse. Also, merge-me is meant for dependabot PRs which should not produce linting issues in general.